[1861/1867] Migrate more games affected by issue #9655
The script in commit 7be7e1ff6e7620f was fixing many of the games broken
by the switch of the order of the order of the 1861/1867 merger round
PostMergerShares and ReduceTokens steps. But it was failing where there
were auto action passes for the merger round.

This adds a helper method to migrate_games.rb to move auto_actions
between actions, and uses this method to move the auto actions from the
last action in the reduce tokens step to the last action in the post
merge shares step.