[1822CA] fix QMOO / Quebec
QMOO's home can be any Quebec city, but actions of other corporations may end up
limiting QMOO's options. Usually when the first yellow tile is laid in Quebec,
QMOO will have to choose its home; now the pending token step will activate
appropriately. This also adds logic to update the home reservation, to ensure
that a specific city is reserved for QMOO if necessary.

Adds a function to the base class to hide hex reservations for corporations, to
declutter the Quebec tile a bit when QMOO is selecting its home.

Adds loc-specific rendering for destinations, for ICR destination on Quebec.