[1868WY] fix LHP assignment when phase 5 is triggered by export
Export means the assignment happens in a stock round instead of an operating
round. The Assign step has a couple of methods implemented so it can render
without error in the stock round, and a newly created `selected_company` method
so the LHP is automatically selected without the player choosing it.

Because the LHP is the active entity, it needs methods from the ShareHolder
module which are called on the entities when rendering a stock round.

Fixes #9714