Initial change to resolve price not dropping on share sales which transfer presidency.  This was caused by the shares not being put into the sell_queue because the presidency had been transferred over at that point, and change_price not being called unless the bundle and the president were the same.
This was resolved by relying on the 1870 implementation and matching
that behavior.  I did not find any rules differences between 1850
and 1870 in this area, and the code for change_price was identical
except @game.stock_market vs. @stock_market type differences.

sell_shares_and_change_price is now identical between 1850 and 1870.