[1822CA] fix getting stuck on track step after a big city upgrades to gray
Fixes issue where after a "big city" is upgraded to gray, and a major owns the
private company that gives free upgrades to that city, the major can get
partially stuck on the track step because of the private's track-laying
ability. The title line shows "upgrade track" as the current step, but the UI
indicates it is time to run routes, and the `available_hexes` don't match the
hexes available to run to. By closing the private company earlier in the
process, the game will no longer return the track step as the current step.

Technically in RAW the big city upgraders never close, but since their income is
$0 and they serve no function once their city is gray, it is simpler to close
them.