Create a queue of corporations to close to fix consecutive closures
Fixes #9894, where, in 1846 two consecutive corporations were closed after all
of their operating steps were automatically skipped due to insufficient cash to
take any actions.

The first corporation (ERIE) was closed correctly, but the second (PRR) was
not. In the `action_processed()` logic [[1][1]] to close companies at $0, PRR
happens to be checked before ERIE; as part of the process for closing ERIE, PRR
starts operating and also has its steps skipped [[2][2]] [[3][3]] and needs to
be closed, but since it was already checked for closure it remains open after
the step that should have closed it.

The iteration in `action_processed()` could be changed to use operating order,
but in cases like this one, since ERIE's price drops before PRR operates, the
operating order check would still have PRR before ERIE. Using `@round.entities`
could work (using the operating order that was computed at the beginning of the
round), but then this function would need more logic to work in stock rounds,
and it gets more complicated should this ever occur across different rounds.

Using Ruby's ordered hash as a queue avoids needing any logic to deal with what
the current round is.

[1]: https://github.com/tobymao/18xx/blob/bcf2e41db61d1e719d6870d2beff6d1772c9a4c2/lib/engine/game/base.rb#L2715
[2]: https://github.com/tobymao/18xx/blob/bcf2e41db61d1e719d6870d2beff6d1772c9a4c2/lib/engine/game/base.rb#L1762
[3]: https://github.com/tobymao/18xx/blob/bcf2e41db61d1e719d6870d2beff6d1772c9a4c2/lib/engine/round/operating.rb#L86