[1858] Make sure home token from private goes to correct city
An improved fix for tobymao#9858.

The issue with a token being placed in the wrong Madrid city can occur
both when a private railway company is acquired by a public company, and
also when a public company is formed from a private railway company.

This fixes the latter, ensuring that the home token can only be placed
in a city where the private railway company held a reservation.