[1858] Clear round.current_actions after an auction
`@round.current_actions` was not being cleared when there was a private
railway company auction. This could lead to the wrong player being given
the priority deal if the following sequence occurred:
1. There was a sell or buy shares action.
2. There were one or more auctions.
3. Everyone passes.

After the auction the `{Sell,Buy}Shares` action was still in
`@round.current_actions` when the active player became the person to the
right of the one who started the auction. This means that the priority
deal would be one place to the right of where it should have been.

This is fixed by clearing `current_actions` in `win_bid`.

Fixes tobymao#9968.