[18Ardennes] Allow trains received from a minor to be discarded
When a minor is exchanged for a share in an existing major corporation,
then its trains can be either transferred or discarded to the bank pool.

This commit also adds a bit of code preparing for allowing station
tokens and fort tokens to be declined.

There's a FIXME in this commit as it doesn't handle the case correctly
where the major and the minor combined have more trains than the current
train limit. In this case the major's existing trains can also be
discarded to stay within limit, but not more than needed to be at limit.