[18Ardennes] Discarding trains when over limit
Better handling of trains when merging a minor puts a major over the
train limit. This is now done as a two-step process. First the major has
the option to decline trains from the minor. Then, if the major is still
over the train limit, they must discard trains (either their original
trains or ones from the minor) to take them back down to the train
limit.