[18Ardennes] Allow tokens from minors to be declined
If a minor is exchanged for a share of a major, then the major's
director is allowed to decline any station tokens.

Add a `DeclineTokens` step (based on `Engine::Step::ReduceTokens`) that
allows the player to select any of the minor's tokens that they don't
want to keep. If these are in port or mine cities then these tokens are
replaced by a neutral port or mine token.