[1822/1822CA] generalize "minor 14" home; [core] fix cheater tokens on tile upgrade
M13 in 1822CA with Toronto works like 1822's M14 and London (it effectively adds
a permanent extra station slot), but London always has 6 cities that normally
have 1 slot each; the Toronto tiles start with 2 separate cities that join in
brown and vary in size (1-4 slots).

The `extra_tokens` and `extra_slot` functionality on `Engine::Part::City` is not
used because the slot needs to persist even if the minor's home token is picked
up and not replaced by a token from the major when it is acquired.

The cheater token fix changes how tokens are moved from the old tile to the new
tile when a tile is upgraded. With the old approach, it was done by iterating
over each city on the old tile, moving tokens city-by-city. The new approach
first gathers all of the tokens, mapped to their new city, sorts them so that
cheater tokens are at the end of the list, and then places the tokens onto the
new city. This ensures that old cheater tokens will be the cheater tokens on the
new city, if the new city doesn't have enough normal slots for all of the
tokens.

M13 is the cheater token in yellow and green, but it was being placed onto the
brown tile before GT's home token, so with tile #T4 (only 2 normal slots), M13's
and M12's tokens were first filled up the normal slots, and then there was no
room for GT's home token. The change ensures that M12 and GT are placed on the
new city before M13 is.

Additionally:

* clean up 1822 dead code, missed in #9519
* update 1822CA fixtures descriptions (missed doing this when fixture 3 was
  added)
* update constants in other 1822-family games for `MINOR_14_ID` references

Fixes #10226