[core/1822CA] Refactor how tokens are moved when upgrading mult-city tiles
`hex.rb` - Refactor moving tokens to new tile to fix bug when multiple cities on
the old tile are being combined into one city on the new tile, and there was a
cheater token in one of the old cities.

Fixes #10226, where M13's home in Toronto (1822CA) is a cheater token but should
not take up added token slots when combined with the other city when a brown T
tile is laid.

M14 in 1822 (and therefore M13 in 1822CA) is implemented as a cheater token,
which normally use up a new token slot when one is added to their city. However,
these tokens do not do that, instead adding a permanent extra slot that could be
vacated if the major chooses so when acquiring the minor. They are not
implemented as `extra_tokens` as those created slots cannot persist.

On the physical board, these extra slots are represented on rondels drawn
offboard, but near the London/Toronto hexes. Perhaps further refactoring could
be done to more robustly implement the rondel functionality and distinguish it
from cheater tokens.

----

Validation against the full database passed.