[1822CA] keep the name "5P" for the special private company 5-train
Renaming it to "5" in this way causes a mismatch with the variant that gets
created when the train object is initialized. Updating the string in the variant
could be done too, but it would be clunky and isn't worth it.

The other 22s name this train "5P", I tried out "5" since once the company
closes and it's actually a train, it works just like any other 5-train.

Fixes #10238