[18Ardennes] Allow forts from minors to be declined
According to the 18Ardennes rules, a major company's director is allowed
to decline the fort tokens from a minor that is swapped for a share.
There's probably never a reason why the major wouldn't want these forts,
but this is a rule, so we'd better make it possible.

This adds a `DeclineForts`, a subclass of `Engine::Step::Choose` that
prompts for the number of fort tokens to be kept.