[18Ardennes] Close minor after exchanged for share
After a minor was exchanged for a share in a major company, it was not
being closed and could still operate in the following operating round.
Make sure that it is closed.

Also, on a minor exchange, only show the hexes with the minor's tokens
as active in the DeclineTokens step. The default behaviour, inherited
from ReduceTokens, was showing hexes with either the minor's or the
major's tokens in.