[18Ardennes] Start on corporations growing up
This is the start of the implementation of the operating round action to
convert a 5-share corporation to a 10-share corporation.

This gives the button to allow the corporation to grow up, and the
director is then allowed to purchase extra shares for cash.

Not yet implemented is exchanging minor companies for shares, or letting
the director go over the overall certificate limit as part of the
conversion process.