[1858] Delete reservation when home token is placed
Fixes tobymao#10442.

An extra slot was being created when Madrid was upgrade to green. The
sequence of events that triggered this was:

1. M&A is used to form AVT in stock round 3.
2. AVT places its home token in M&A's slot in the NE Madrid city.
3. In OR 3.1 RP operates before AVT.
4. RP upgrades Madrid to green.

At this point M&A hasn't yet closed and the code for moving tokens and
reservations sees both AVT's token and M&A's reservation in the NE
Madrid city, and helpfully creates a second slot to accommodate both of
these.

The fix is to remove M&A's reservation after AVT places its home token.