Prevent spurious exchange buttons appearing in stock round
Bugs tobymao#10537 and tobymao#10563 were reports of spurious exchange
buttons appearing in stock rounds, in 1871 and 1858.

These buttons were appearing when the active player owned a private
company that could be exchanged for a share in a corporation. As well as
the button to exchange the private for a share, one or more extra
buttons were appearing, offering to exchange corporations for a share.

This was caused by commit bc6fa7d, which added `@game.corporations` to
the entities that are checked for exchange abilities. This was needed
for 18Ardennes, which has corporations with exchange abilities.

Neither 1858 or 1871 have corporations with exchange abilities, but
`Corporation.all_abilities` will also pick up any abilities owned by the
corporation's president which have an owning_player-type timing. This
caused these corporations to appear as an option to be exchanged.

This is fixed by removing `@game.corporations` from the default entities
that are checked for exchange abilities. To allow 18Ardennes to still
check those, a new method `@game.exchange_entities` is added which
defines which entities are examined for the exchange abilities. The
default implementation returns all companies and minors, and so reverts
to the behaviour before commit bc6fa7d. 18Ardennes has its own version
of this method added, which checks the minor-type corporations.