[18Ardennes] Redo finding minors eligible to start a public company
The existing method for checking which minors could be used to start a
public company was not working where they had a token in Paris. Tokens
in the western Paris city can only be used to start N and tokens in the
eastern city can only be used to start E. The old code was allowing
either N or E to be started using any Parisian token.