[18Ardennes] Require approval for exchanging minors
If a minor is being exchanged for a public company's treasury share,
then this requires the approval of the public company's president,
unless the public company has just formed in the current stock round.

This reuses the approval mechanism from 1858.

Still to be implemented is only allowing exchanges for a market share if
the public company's president has previously denied a request to
exchange for a treasury share.