[18Ardennes] Add auto-action for pass in stock round
In `Step::BuySellParSharesCompanies.actions` there is no check for a
legal connection between a minor company and a public company, to avoid
recalculating the game graph when loading a game. This means that it is
possible for `actions` to return a `buy_shares` action when there is no
legal exchange possible.

Make sure that a player isn't given a stock turn when there is no
possible action other than passing by adding a pass auto-action if
there are no legal exchange or buy actions.