[18Ardennes] Better message for not being able to place bid
Improved message when skipping a player in the public company auction
round if they do not have any available minors to place a bid. This can
either because they do not own any minors or all of them have been
already pledged in bids. The message was just suggesting the first of
these, so make it more general.