[18Ardennes] Selling the GL minor ends a stock turn
v3 of allowing the GL to be sold before starting a public company. The
previous version was overriding `can_sell_any_companies?` with a version
that did not check `bought?`. This meant that if a player did a buy
stock round action their turn was not ending if they owned the GL.