[18Ardennes] No stock round after the initial auction
There's no point having a stock round after the auction of the minor
companies. The rules aren't clear on whether there should be a stock
round between the auction and the first operating round, but in three-
and five-player games nobody can do anything apart from pass, and in a
four-player game the only legal move is to sell the GL minor, and that's
utterly stupid. So go straight to the operating round.