[1822PNW] create placeholder company to occupy empty bid boxes
When an associated Minor is removed from a bid box because its associated Major
is started directly, a placeholder company is used to fill the bidbox, and is
cleaned up at the end of the stock round.

Fixes #10525 and #10609, where an associated Minor was removed from bid box 1,
and the Minor in bid box 2 received no bids and caused a train to export as if
it had been in bid box 1.