[Code] Minor's abilities can be used in stock round
Commit 67572e8 had broken 1858. This commit introduced changes to
ability timing, introducing a check that an ability with timing
`owning_player_sr_turn` was either owned by a player or a player's
company.

1858 includes abilities that are owned by minor entities which were
failing this check. This meant that it was not possible to exchange a
private railway company (minor+company pair) for a share in a public
comany (corporation).

Fix this by changing the check to work for minors too.

Fixes issue tobymao#10813.