[Core/1822MX] Simplify base implementation of Game.can_dump?
The implementation of `Game.can_dump?` introduced by tobymao#10907 broke
the (pre-alpha) implementation of 18Ardennes. It caused a 'maximum stack
size exceeded' error. This is because there is a step in 18Ardenees
where `actions` calls `@game.liquidity`, and so the call to
`active_step` in this new `can_dump?` function is causing an infinite
loop.

To fix this, and avoid the same error being caused in future, this
commit makes the base class implementation of `can_dump?` as simple as
possible, just calling `Bundle.can_dump?`. The version of the method
which checks `active_step` is moved into the 1822MX game code.