[1858] Fix graph error where cities on home hex have merged
Private railways can start in specific cities on a hex with multiple
cities and are only allowed to track routes from that city. If the
cities merge before the private closes then a nil key was being added to
the `home_hex_nodes` hash, which was causing an error when trying to
walk the graph.

The cities only ever merge into a single city, so we can trace the route
from the only city on the tile in this case.