[18Ardennes] Fix error on tokening multi-slot city
An error was occurring when attempting to token the brown Ruhr tile
(3-slot city). There was already a token in slot #0. If the player
clicks on slot #1 when placing a token this worked correctly, but
clicking on slot #2 errored, as the game was then attempting to change
the logo of the token in slot #2, but the token had gone to slot #1.

Fixes ollybh#133.