[18Ardennes] Choose player to operate companies in receivership
If a minor or public company is in receivership and it owns a train then
someone needs to select its routes. To make it easier for players,
choose the player who has just operated a company, or the next one to do
this if none have yet operated in the current operating round.