[1824] 10 trains should not rust any g trains
There is an error in the English rule book, section VII.14, which claims
that 10 trains rust any remaining 1g, 2g, and 3g trains.
But this is a translation error.

The designer (Lonny) has also explicitly pointed this out in a forum
thread about this. A link to this is present on the 1824 wiki page.