[1858Switzerland] Add entities, map, tiles, icons and logos
1858 Switzerland is a new prototype from Ian D Wilson. As the name
suggests, it is a game based on 1858 set in Switzerland.

This adds definitions of:
 - The public companies (majors, 5-/10-share corporations).
 - The private railway companies (implemented as a pair of a private and
   a minor).
 - The trains and phases.
 - The map.
 - The tiles.

There's almost no actual code in this, just enough to override some of
the definitions for trains and phases inherited from the base 1858 code.

It also adds SVG icons and logos for the companies, plus a few icons for
the map.