Suppress profile links for dummy players
Several games have dummy players:

- The Union Bank in 1871.
- Vaclav in the two-player version of 18CZ.
- A W Edelswärd in the two-player version of 18SJ.

These are all represented by player objects with an ID of -1.

These entities are shown on the entity page (and other places) where a
link to the player profile pages is included. Clicking on these links
gives a 404 error, as these links are not valid.

This pull request suppresses the links to the profile pages for players
with negative user IDs.

Fixes tobymao#11124.