Move the auto-router path timeout check to an inner loop
The auto-router's path timeout is supposed to stop too much time being
spent on building up the set of possible routes for trains. It was being
checked once in each iteration of possible starting nodes for routes.

However, in particularly degenerate cases this meant that far too long
was elapsing between timeout checks: issue tobymao#11252 had more than
ten hours pass before the first and second timeout checks.

This moves the timeout check to an inner loop, so it is checked more
frequently.