[1822/core] fix 1822 tax haven bug
1822's `company_status_str` (called by views) was mutating the value of the
private company entity, and there was a substraction operation in place to try
to copmensate for that, but that caused #11040 by subtracting the tax haven's
value for other players.

This has been simplified with the following approach:

- when displaying the company value, call a new wrapper method in the `Game::Base` class
- in 1822, override that new method only for the Tax Haven

Now the Tax Haven private company entity will always have a value of $0/£0, but the
value of the Tax Haven "player" entity is handed off to display methods.

Fixes #11040