[1858 Switzerland] Avoid modifying 1858's train definitions
tobymao#11372 was caused by modifying 1858's train definitions without a
making a deep copy. If a game of 1858 is loaded on the server after a
game of 1858 Switzerland then it is was seeing the version of the definitions
after these modifications, causing errors as 1858 does not have a method
for handling the `blue_privates_available` event.

Making a deep copy of the train definitions prevents this error.

This commit also makes a deep copy of the phase definitions. No problems
have been observed with this, but making the copy is safer.

Fixes tobymao#11372.