[Core] Don't modify TRAINS definition when creating train variants
The `variants` parameter passed to `Engine::Train.init_variants` is (in
the default implementations) an array of train variant definitions that
is part of the TRAINS constant. The code inside `init_trains` was, for
each train, creating a new variant for the normal type of train and
prepending it to the variant array. This was modifying the array inside
the TRAINS constant, making the list of variants steadily longer.

This wasn't much of a problem in most cases as the extra items in the
`variants` array were all duplicates and we only extra a single item for
each variant name. But when games are loaded on the server in quick
succession the TRAINS constant is shared between a game title and
(potentially) any titles that inherit from that title. This was causing
bug tobymao#11416, where 1858 Switzerland was attempting to modify the
price of a variant, and was expecting there to only be a single item in
the `variants` array for the type of train.

This commit changes the `init_variants` method to not make any changes
to the array it is passed, instead copying the contents to a new array.