[1858] Rewrite Graph::home_hex_nodes method
`home_hex_nodes` is used in 1858 to find the nodes in the home hexes of
the private railway companies.

For plain track hexes it creates some dummy nodes, and for hexes with
towns and most cities it returns those revenue centres.

The other case is for Madrid in 1858, which has multiple cities on the
same hex. Each private railway in Madrid can only trace a route from its
own city circle, not from anywhere in the Madrid hex. This was done by
hardcoding the city index of the privates' homes, and working from
there, with checks to see if the Madrid cities have merged.

This works for 1858 because there is only one possible orientation of
the green Madrid tile. 1858 India follows the same principle for the
Calcutta and Delhi hexes, but both of these have two possible
orientations of the green tiles, and in one the city indexes do not
match the yellow tile. This meant that the routes for the privates could
be traced from the wrong city circles.

This changes the `home_hex_nodes` method to, instead of looking for a
city index, find the city that the private railway has a reservation
for, and trace the route for this. This should be more reliable and
removes the need to specify the starting city for the minor entity
component of the private railways.