[1858/1858 Switzerland] Move phase 4 train rusting trigger to a constant
In 1858 (Iberia) the phase 4 trains (6H/3M) rust when a certain number
of grey trains (7E/6M/5D) have been bought. This had been hardcoded in
the game's `setup` method, this commit moves this to a constant.

This is to make this number easier to override in 1858 India which has a
higher limit.