[1858/1858 Switzerlad] Use company types for private railway batches
In the 1858 family of games the private railway companies become
available in batches: first the yellow privates then the green privates,
and (in 1858 Switzerland only) the blue ones last.

This had been implemented by looking at the `color` attributes of the
company objects. This changes this by giving each company a type
attribute corresponding to its batch and using those instead.