[Core] Check for existence of `can_finance?` method
The `can_finance?` method was added for 1837 in commit d08cc8435. It's
used to check whether the bank will step in to pay the cost of a train.

This has broken 1877 Venezuela. There is a default implementation of
`can_finance?` added to the Step::BuyTrain class, but 1877 Venezuela
allows trains to be bought in the BuySellParShares step, which did not
have this method.

This commit tests for the existence of this method before calling it.

Fixes tobymao#11508.