[1844] Fix error from nil ability description
When the first 6/6H train is bought, the destinations for pre-SBB and
large historical companies are removed. This involves removing the
destination abilities for these companies, but in tobymao#11591 this was
causing an error.

The test for a destination ability is done for looking at the ability's
description. BLS has, as well as the destination ability, an assign
hexes ability. That does not have a description set, and so an
`undefined method 'start_with?' for nil` was being thrown. This is fixed
by using the safe navigation operator.

Fixes tobymao#11591.