[Core] Don't assume all operating entities have trains
tobymao#11586 made some changes to the setup method of operating rounds.
This has caused an error in some games of 1868 Wyoming, where P6c is one
of the entities in the operating round. P6c is a company and does not
have a trains attribute.

This adds a guard cause, to avoid errors like this.

Fixes tobymao#11594.