[1858] Clear pass flag when bidding for a private
A player's pass status was not being cleared when a bid for a private
railway company was placed. This meant that it was possible for the
initial stock round to end prematurely – if a player had passed on an
auction earlier in the round but bid on a later one they were still
marked as having passed, meaning the round could end if all other
players had passed.

Fixes tobymao#11446.