[1858 India] Fixes for locomotive works privates
The locomotive works privates don't behave the same as company objects
in 1858: they pay income at the start of operating rounds, and don't add
their income to the train revenue, or close automatically when owned by
a corporation.

Override the `G1858::private_railway?` method to make these company
objects behave like standard 18xx companies. Also use this method in
other places where we need to work out if the company is for a private
railway or a locomotive works.