[Core/1868WY] Add method for checking if a hex token can be removed
The `remove_hex_token` action is only currently used by 1868WY. There is
a check built into the view class to make sure that the token is owned
by the currently operating entity.

This moves this check to the step class, and calls it from the view
class. This is to allow other games (1858 India) to implement different
checks as to whether a hex token can be removed.