[1858] Refactor: extract `trainless?` method
This is to make implementing 1858 India easier. That has a type of train
(mail trains) which does not stop a company entering emergency money
raising if they only own this type of train.

Add a `trainless?(corporation)` method to replace checks to
`corporation.trains.empty?`. This implementation just does the
`trains.empty?` check, but 1858 India will be able to override this
method with a version that checks for ownership of a non-mail train.