[1858 India] Allow EMR share issuing if corp only owns a mail train
A mail train isn't a normal train, and if a public company only owns a
mail train then it is still allowed to issue shares to allow it to
purchase a train.

Also remove the overriden `must_buy_train?` method. As this is an 1858
derivative, a public company is never obliged to own a train.