[1862] Fix autopathing issues
In 1862, the autopather can't find paths which don't include the home token.
This is because the check for the home token is in `check_overlap` when it
should be in `check_other`. `check_overlap` is intended to be used to make sure
trains don't reuse tracks whereas `check_other` is for checks that require the
full set of routes to check the validity of. The key is that the autopather
doesn't call `check_other` when calculating the set of all possible routes per
train.