[1867] Refactor: extract `merger_round` and `new_merger_round` methods
In the 1867 code the steps that make up a merger round were embedded in
the `next_round!` method. Refactor this method to extract
`new_merger_round` and `merger_round` methods, to make it easier to
override merger round steps.

This is needed for the 1807 code. In that game, as well as 1861/1867
mergers of minors into public companies, you have public companies
merging to form systems.