[18ESP] Updated rules URLs
The link for the game rules had broken. It looks like the BGG file page
for the draft rules has been deleted now that a final version of the
rules is available.

This updates the rules link to point to a BGG page for the final version
of the Lonny Games rules. It also adds a link to the Playbook file page.