[1867] Refactor: extract `fix_token_count!` method
When minors merge to form a major (public company) in 1861/1867, the
merger code can remove all but two of their tokens. There is code to add
back the third token.

This code is duplicated in a couple of places. This commit extracts this
to a single method, removing this duplication, and also making it easier
to override this method when different behaviour is required.

1807 is an 1681 derivative, but the large companies (public companies
and systems) have 4, 8 or 12 tokens each, and the token cost is always
£20 per hex, not the 40₽/$40 per hex for the last token you have in
1861/1867. Extracting this method makes it easier to implement the
correct behaviour in 1807.