[1807] Correct token calculation cost from London tokens
The previous fix for ollybh#190 was incomplete. It had fixed the
calculation of token cost when placing a token in London, but had not
considered that you also need to look at the cost of placing a token
outside London when a company has a token in London.

This commit changes the calculation so that all token costs will be
calculated using the location of London on the main map (hex U13), not
the zoomed in London hexes in the North Sea.