[1861/1867] Don't transfer zero-cost tokens from minors to majors
The fix for tobymao#8589 in PR tobymao#8689 was incomplete. It did not
handle the case of two minor companies with tokens in the same city
being merged to form a new major (public company). When this type of
merger was being done, one of the two minor companies' token was being
removed by the player and was returned to the minor's charter. The call
to `move_tokens_to_surviving` from
`Step::ReduceTokens.process_reduce_tokens` was then moving this token
from the minor to the major, where it was keeping the zero cost it had
in the minor company.

This prevents this from happening by clearing all unplaced tokens from
minor companies at the start of `move_tokens_to_surviving`. 1861/1867
does not need to move unplaced tokens around, the major companies
already have as many tokens as they are allowed.

Fixes tobymao#11750.