[Core] Allow games to have tests for valid mergers/exchanges
In the merger round view allow game steps to expose tests that can be
called to determine whether a valid selection has been made for a
corporation to take part in a merger or an exchange.

If the selection is not valid then the button is disabled and no action
assigned to it.

This defaults to the current behvaiour, where selecting any corporation
is enough to enable the button. If the step implements a
`merge_allowed?` method then this will be called with the active
corporation and the select corporation passed as parameters. The button
will only be enabled if this method returns true.

There is a similar call to the `exchange_allowed?` method for the
exchange button. This one is not wrapped in a `respond_to?` test as this
will only be called for 1807 and that does implement this method.