[1858 India] Show mine/oil/port token cost on the timeline
The cost of mine, oil and port tokens was only visible in the
description on the page for the collect tokens step. To make this
information available at all time, this adds these costs to the timeline
on the info page. This also lists the hexes which still have these
tokens.