[1858 India/Core] Show loco work privates on left in buy trains step
The locomotive works private companies can only be purchased during the
buy trains step. The default rendering of this step puts private
companies underneath the game map. 1858 India has a large map, so
there's a high chance that these privates will not be visible unless the
player scrolls the page downwards.

To make these privates more visible, and reduce the chance of players
not realising that purchasing them is an option, this moves the
companies to the left side of the page, where they will be rendered with
trains that can be purchased and the company card.

Fixes ollybh#164.