autorouter: Fix timeout/early exit code
We weren't actually stopping running when this.router.running was false.
This was noticeable when navigating away from the turn while autorouting.

Also fix exception code since the logs would just dump a bare nil as the
exception backtrace and forgot to set the routes to the best known routes.