[1822] Don't remove corporations' home tokens
When multiple cities on a tile merge, a corporation which had tokens in
the different cities will have all but one returned to its charter.

Which token was returned depended on the order of the cities on the
tile, with the tokens in the higher index cities removed. It was
possible for a corporation's initial token to be removed by this
algorithm. This was breaking the calculation of destination bonuses, as
the check for a run between a coporation's home and destination is made
by looking at its first and destination tokens.

This sorts the duplicated tokens by their index in the corporation's
`tokens` array. This ensures that the home token, the first token in
this array, will not be removed.

Fixes tobymao#11845.