[1858 India] Changed cross-grading rules
It is now allowed to change gauge on a city tile when you are changing
the gauge on a single path, and maintaining the same gauge on the
others.

This is to allow an upgrade from a yellow city tile with one broad gauge
path and one narrow gauge path, to either an all broad gauge city tile,
or an all narrow gauge tile.

Fixes ollybh#207.