Make par_chart overridable
If a game implements par_chart, and a derived game don't, it should be possible
to override the behavior. This can be done by implementing par_chart and let it return nil.

This should not affect existing implementations that uses par_chart (1837 and 1880) but
helps games that uses these as base (1824).