[Core] Tweak to algorithm for restoring previous routes
When an old route is being loaded, either for route history or to
populate the route selector, we attempt to map the route onto upgraded
tiles.

Part of this process is done in `Route.find_matching_chains`. If there
are multiple candidate routes between tiles A1→B2→C3 it will look for
nodes that were on the previous route. If the previous route visited
nodes A1-0, B2-1 and C3-1 it will prefer a candidate route that includes
all three of those nodes. But if no candidates include all three nodes
then the first candidate route is returned.

This changes the algorithm to return the candidate route which matches
the most nodes. So if there are two candidate routes where the first
visits nodes [A1-0, B2-0, C3-0] and the second visits [A1-0, B2-1, C3-0]
it will prefer the second candidate as it matches two nodes wheras the
first matches just one.

Fixes tobymao#11872.