[1822PNW] Port tiles must connect to large track spikes
Water hexes can have both short and long track spikes. The short ones
(a path with a terminal attibute of value '2') are to allow tiles
adjacent to the water to be upgraded. The long track spikes (a path with
a terminal attribute of value '1') show the legal orientations of the
port tiles.

The existing code for checking the port tile orientation is only
checking whether the track on the port tile replaces one of the track
connections on the water tile. It does not check the type of track, so
it is possible to lay the tile in an illegal orientation, replacing a
short track spike instead of a long one.

Fixes tobymao#11927.